requirements.txt
anthropic==0.25.6
beautifulsoup4==4.12.3
openai==1.23.2
protobuf==3.20.3
python-dotenv==1.0.1
streamlit==1.32.0
google-generativeai==0.5.0